Changes for moving binary function
