+++ src/modules/string.cpp
@@ -10,7 +10,7 @@
 #include "common/algo_helper.h"
 #include "common/format_helper.h"
 #include "faker-cxx/helper.h"
-#include "faker-cxx/number.h"
+// #include "faker-cxx/number.h"
 #include "faker-cxx/types/hex.h"
 #include "string_data.h"
 
@@ -354,66 +354,6 @@ std::string hexadecimal(GuaranteeMap&& guarantee, unsigned int length, HexCasing
     const auto& hexadecimalPrefix = hexPrefixToStringMapping.at(prefix);
 
     return hexadecimalPrefix + generateStringWithGuarantee(guarantee, targetCharacters, length);
-}
-
-std::string binary(int length)
-{
-    if (length < 0)
-    {
-        throw std::invalid_argument("The length of a binary number cannot be negative");
-    }
-
-    std::string binaryNumber;
-
-    for (int i = 0; i < length; ++i)
-    {
-        binaryNumber += static_cast<char>(number::integer(1));
-    }
-
-    return "0b" + binaryNumber;
-}
-
-std::string binary(int min, int max)
-{
-    if (min > max)
-    {
-        throw std::invalid_argument("min cannot be greater than max");
-    }
-
-    if (min < 0 || max < 0)
-    {
-        throw std::invalid_argument("The output binary string cannot be negative");
-    }
-
-    int num = number::integer<int>(min, max);
-
-    if (num == 0)
-    {
-        return "0b0";
-    }
-
-    std::string output;
-
-    while (num > 0)
-    {
-        int remainder = num % 2;
-        output = std::to_string(remainder) + output;
-        num /= 2;
-    }
-
-    return "0b" + output;
-}
-
-std::string binary(GuaranteeMap&& guarantee, unsigned int length)
-{
-    std::set<char> targetCharacters{'0', '1'};
-
-    if (!isValidGuarantee(guarantee, targetCharacters, length))
-    {
-        throw std::invalid_argument{"Invalid guarantee."};
-    }
-
-    return "0b" + generateStringWithGuarantee(guarantee, targetCharacters, length);
 }
 
 std::string octal(unsigned int length)
